+++ release/changes.md
@@ -1,1 +1,2 @@
-- [NEW] TBD
+- [FIX] Rename `AI Agent` tag to `AI`
+- [FIX] Rename `AI Agent` value key to `AI agent`